refactor: remove legacy version of providers
