--- wallets/provider-coinbase/src/legacy/index.ts
@@ -1,183 +1,0 @@
-import type { Provider } from '../utils.js';
-import type {
-  CanEagerConnect,
-  CanSwitchNetwork,
-  Connect,
-  ProviderConnectResult,
-  Subscribe,
-  SwitchNetwork,
-  WalletInfo,
-} from '@rango-dev/wallets-shared';
-import type { BlockchainMeta, SignerFactory } from 'rango-types';
-
-import {
-  type LegacyProviderInterface,
-  LegacyNetworks as Networks,
-} from '@rango-dev/wallets-core/legacy';
-import {
-  canEagerlyConnectToEvm,
-  canSwitchNetworkToEvm,
-  chooseInstance,
-  getEvmAccounts,
-  switchNetworkForEvm,
-  WalletTypes,
-} from '@rango-dev/wallets-shared';
-import {
-  evmBlockchains,
-  isEvmBlockchain,
-  isSolanaBlockchain,
-  solanaBlockchain,
-} from 'rango-types';
-
-import signer from '../signer.js';
-import {
-  coinbase as coinbase_instance,
-  getSolanaAccounts,
-  solanaCoinbase,
-} from '../utils.js';
-
-const WALLET = WalletTypes.COINBASE;
-
-export const config = {
-  type: WALLET,
-  defaultNetwork: Networks.ETHEREUM,
-};
-
-export const getInstance = coinbase_instance;
-export const connect: Connect = async ({ instance, meta }) => {
-  /*
-   * Note: We need to get `chainId` here, because for the first time
-   * after opening the browser, wallet is locked, and don't give us accounts and chainId
-   * on `check` phase, so `network` will be null. For this case we need to get chainId
-   * whenever we are requesting accounts.
-   */
-  const evm_instance = chooseInstance(instance, meta, Networks.ETHEREUM);
-  const results: ProviderConnectResult[] = [];
-
-  if (evm_instance) {
-    const evm = await getEvmAccounts(evm_instance);
-    results.push(evm);
-  }
-
-  const solanaInstance = solanaCoinbase();
-  if (solanaInstance) {
-    const solanaResults = await getSolanaAccounts(solanaInstance);
-    results.push(solanaResults);
-  }
-
-  return results;
-};
-
-export const subscribe: Subscribe = (options) => {
-  const ethInstance = chooseInstance(
-    options.instance,
-    options.meta,
-    Networks.ETHEREUM
-  );
-  const solanaInstance = chooseInstance(
-    options.instance,
-    options.meta,
-    Networks.SOLANA
-  );
-  const { connect, updateAccounts, state, updateChainId, meta } = options;
-  const handleEvmAccountsChanged = (addresses: string[]) => {
-    const eth_chainId = meta
-      .filter(isEvmBlockchain)
-      .find((blockchain) => blockchain.name === Networks.ETHEREUM)?.chainId;
-    if (state.connected) {
-      if (state.network != Networks.ETHEREUM && eth_chainId) {
-        updateChainId(eth_chainId);
-      }
-      updateAccounts(addresses);
-    }
-  };
-
-  const handleSolanaAccountChanged = async (publicKey: string) => {
-    if (state.network != Networks.SOLANA) {
-      updateChainId(meta.filter(isSolanaBlockchain)[0].chainId);
-    }
-    const network = Networks.SOLANA;
-    if (publicKey) {
-      const account = publicKey.toString();
-      updateAccounts([account]);
-    } else {
-      connect(network);
-    }
-  };
-  ethInstance?.on?.('accountsChanged', handleEvmAccountsChanged);
-
-  solanaInstance?.on?.('accountChanged', handleSolanaAccountChanged);
-
-  return () => {
-    ethInstance?.off?.('accountsChanged', handleEvmAccountsChanged);
-
-    solanaInstance?.off?.('accountChanged', handleSolanaAccountChanged);
-  };
-};
-export const switchNetwork: SwitchNetwork = switchNetworkForEvm;
-
-export const canSwitchNetworkTo: CanSwitchNetwork = canSwitchNetworkToEvm;
-
-export const getSigners: (provider: Provider) => Promise<SignerFactory> =
-  signer;
-
-export const canEagerConnect: CanEagerConnect = async ({ instance, meta }) => {
-  const evm_instance = chooseInstance(instance, meta, Networks.ETHEREUM);
-  if (evm_instance) {
-    return canEagerlyConnectToEvm({ instance: evm_instance, meta });
-  }
-  return Promise.resolve(false);
-};
-
-export const getWalletInfo: (allBlockChains: BlockchainMeta[]) => WalletInfo = (
-  allBlockChains
-) => {
-  const evms = evmBlockchains(allBlockChains);
-  const solana = solanaBlockchain(allBlockChains);
-  return {
-    name: 'Coinbase',
-    img: 'https://raw.githubusercontent.com/rango-exchange/assets/main/wallets/coinbase/icon.svg',
-    installLink: {
-      CHROME:
-        'https://chrome.google.com/webstore/detail/coinbase-wallet-extension/hnfanknocfeofbddgcijnmhnfnkdnaad',
-      BRAVE:
-        'https://chrome.google.com/webstore/detail/coinbase-wallet-extension/hnfanknocfeofbddgcijnmhnfnkdnaad',
-      DEFAULT: 'https://www.coinbase.com/wallet',
-    },
-    color: '#2a62f5',
-    supportedChains: [...evms, ...solana],
-    needsNamespace: {
-      selection: 'multiple',
-      data: [
-        {
-          label: 'EVM',
-          value: 'EVM',
-          id: 'ETH',
-          getSupportedChains: (allBlockchains: BlockchainMeta[]) =>
-            evmBlockchains(allBlockchains),
-        },
-        {
-          label: 'Solana',
-          value: 'Solana',
-          id: 'SOLANA',
-          getSupportedChains: (allBlockchains: BlockchainMeta[]) =>
-            solanaBlockchain(allBlockchains),
-        },
-      ],
-    },
-  };
-};
-
-const buildLegacyProvider: () => LegacyProviderInterface = () => ({
-  config,
-  getInstance,
-  connect,
-  subscribe,
-  switchNetwork,
-  canSwitchNetworkTo,
-  getSigners,
-  getWalletInfo,
-  canEagerConnect,
-});
-
-export { buildLegacyProvider };
