--- wallets/provider-metamask/src/legacy/index.ts
@@ -1,91 +1,0 @@
-import type { LegacyProviderInterface } from '@rango-dev/wallets-core/legacy';
-import type {
-  CanEagerConnect,
-  CanSwitchNetwork,
-  Connect,
-  Subscribe,
-  SwitchNetwork,
-  WalletInfo,
-} from '@rango-dev/wallets-shared';
-import type { BlockchainMeta, SignerFactory } from 'rango-types';
-
-import {
-  canEagerlyConnectToEvm,
-  canSwitchNetworkToEvm,
-  getEvmAccounts,
-  subscribeToEvm,
-  switchNetworkForEvm,
-  WalletTypes,
-} from '@rango-dev/wallets-shared';
-import { evmBlockchains } from 'rango-types';
-
-import signer from '../signer.js';
-import { metamask as metamask_instance } from '../utils.js';
-
-const WALLET = WalletTypes.META_MASK;
-
-export const config = {
-  type: WALLET,
-};
-
-export const getInstance = metamask_instance;
-export const connect: Connect = async ({ instance }) => {
-  /*
-   * Note: We need to get `chainId` here, because for the first time
-   * after opening the browser, wallet is locked, and don't give us accounts and chainId
-   * on `check` phase, so `network` will be null. For this case we need to get chainId
-   * whenever we are requesting accounts.
-   */
-  const { accounts, chainId } = await getEvmAccounts(instance);
-
-  return {
-    accounts,
-    chainId,
-  };
-};
-
-export const subscribe: Subscribe = subscribeToEvm;
-
-export const switchNetwork: SwitchNetwork = switchNetworkForEvm;
-
-export const canSwitchNetworkTo: CanSwitchNetwork = canSwitchNetworkToEvm;
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type Provider = any;
-export const getSigners: (provider: Provider) => Promise<SignerFactory> =
-  signer;
-
-export const canEagerConnect: CanEagerConnect = canEagerlyConnectToEvm;
-
-export const getWalletInfo: (allBlockChains: BlockchainMeta[]) => WalletInfo = (
-  allBlockChains
-) => {
-  const evms = evmBlockchains(allBlockChains);
-  return {
-    name: 'MetaMask',
-    img: 'https://raw.githubusercontent.com/rango-exchange/assets/main/wallets/metamask/icon.svg',
-    installLink: {
-      CHROME:
-        'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn?hl=en',
-      BRAVE:
-        'https://chrome.google.com/webstore/detail/metamask/nkbihfbeogaeaoehlefnkodbefgpgknn?hl=en',
-
-      FIREFOX: 'https://addons.mozilla.org/en-US/firefox/addon/ether-metamask',
-      EDGE: 'https://microsoftedge.microsoft.com/addons/detail/metamask/ejbalbakoplchlghecdalmeeeajnimhm?hl=en-US',
-      DEFAULT: 'https://metamask.io/download/',
-    },
-    color: '#dac7ae',
-    supportedChains: evms,
-  };
-};
-const buildLegacyProvider: () => LegacyProviderInterface = () => ({
-  config,
-  getInstance,
-  connect,
-  subscribe,
-  getSigners,
-  getWalletInfo,
-  canSwitchNetworkTo,
-  canEagerConnect,
-});
-
-export { buildLegacyProvider };
