+++ wallets/provider-safepal/src/mod.ts
@@ -1,12 +1,8 @@
 import { defineVersions } from '@hub3js/core/utils';
 
-import { buildLegacyProvider } from './legacy/index.js';
 import { buildProvider } from './provider.js';
 
 const versions = () =>
-  defineVersions()
-    .version('0.0.0', buildLegacyProvider())
-    .version('1.0.0', buildProvider())
-    .build();
+  defineVersions().version('1.0.0', buildProvider()).build();
 
 export { versions };
